+++ src/Symfony/Component/Security/Core/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.2
 ---
 
+ * Add the `IS_AUTHENTICATED_RECENTLY` attribute, granted while the last interactive authentication is within the configured lifetime
  * Deprecate passing more than one Security attribute to `AccessDecisionManager::decide()`, the `$allowMultipleAttributes` argument will be removed in 9.0
  * Allow a list of identifiers in the `$aud` argument of `OAuth2User`, as RFC 7662 §2.2 defines it
  * Allow using wildcards as placeholders in the keys of the `RoleHierarchy` map